+++ djangocms_blog/forms.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 from django import forms
 from django.conf import settings
+from django.contrib.auth import get_user_model
 from django.core.validators import MaxLengthValidator
 from django.utils.functional import cached_property
 from parler.forms import TranslatableModelForm
@@ -10,6 +11,8 @@ from taggit_autosuggest.widgets import TagAutoSuggest
 
 from .models import BlogCategory, BlogConfig, Post
 from .settings import PERMALINK_TYPE_CATEGORY, get_setting
+
+User = get_user_model()
 
 
 class ConfigFormBase(object):
@@ -78,6 +81,15 @@ class LatestEntriesForm(forms.ModelForm):
                 settings.STATIC_URL, 'djangocms_blog_admin.css'
             ),)
         }
+
+
+class AuthorPostsForm(forms.ModelForm):
+    def __init__(self, *args, **kwargs):
+        super(AuthorPostsForm, self).__init__(*args, **kwargs)
+        # apply distinct due to django issue #11707
+        self.fields['authors'].queryset = User.objects.filter(
+            djangocms_blog_post_author__publish=True
+        ).distinct()
 
 
 class PostAdminFormBase(ConfigFormBase, TranslatableModelForm):