+++ HISTORY.rst
@@ -12,6 +12,8 @@ History
 * Set minimum django-parler version to 2.0
 * Fix to respect current locale for category names
 * Reduce the maximum length of slug fields to 752 characters
+* Fix duplicated authors in BlogAuthorPostsPlugin
+* Add BlogAuthorPostsListPlugin to show posts per author
 
 *******************
 1.0.0 (2019-11-04)